# package/openapi/models/Auth.ts
/* tslint:disable */
/* eslint-disable */
/**
 * User and Team Management API
 * API for user and team management with JWT internal
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Auth
 */
export interface Auth {
    /**
     * 
     * @type {string}
     * @memberof Auth
     */
    token?: string;
}

/**
 * Check if a given object implements the Auth interface.
 */
export function instanceOfAuth(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function AuthFromJSON(json: any): Auth {
    return AuthFromJSONTyped(json, false);
}

export function AuthFromJSONTyped(json: any, ignoreDiscriminator: boolean): Auth {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'token': !exists(json, 'token') ? undefined : json['token'],
    };
}

export function AuthToJSON(value?: Auth | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'token': value.token,
    };
}