# package/openapi/models/SigninRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * User and Team Management API
 * API for user and team management with JWT internal
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SigninRequest
 */
export interface SigninRequest {
    /**
     * User's username
     * @type {string}
     * @memberof SigninRequest
     */
    email: string;
    /**
     * User's passwor
     * @type {string}
     * @memberof SigninRequest
     */
    password: string;
}

/**
 * Check if a given object implements the SigninRequest interface.
 */
export function instanceOfSigninRequest(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "email" in value;
    isInstance = isInstance && "password" in value;

    return isInstance;
}

export function SigninRequestFromJSON(json: any): SigninRequest {
    return SigninRequestFromJSONTyped(json, false);
}

export function SigninRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): SigninRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'email': json['email'],
        'password': json['password'],
    };
}

export function SigninRequestToJSON(value?: SigninRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'email': value.email,
        'password': value.password,
    };
}